Minor review updates for file length checks.
git-svn-id: svn://svn.code.sf.net/p/passwordsafe/code/trunk/pwsafe/pwsafe@2140 1f79f812-37fb-46fe-a122-30589dd2bf55

+++ corelib/Util.h
@@ -54,7 +54,7 @@ extern size_t _readcbc(FILE *fp, unsigned char* &buffer,
                        unsigned char &type, Fish *Algorithm,
                        unsigned char* cbcbuffer,
                        const unsigned char *TERMINAL_BLOCK = NULL, 
-                       long file_len = 0);
+                       size_t file_len = 0);
 extern size_t _writecbc(FILE *fp, const unsigned char* buffer, int length,
                         unsigned char type, Fish *Algorithm,
                         unsigned char* cbcbuffer);
